phase-2/src/btc_wrapper.py
#!/usr/bin/env python3

import os
import sys
import json
import logging
import asyncio
from typing import Dict, List, Optional
from datetime import datetime
from web3 import Web3
from eth_account import Account
from bitcoin.rpc import RawProxy
from cryptography.fernet import Fernet

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('logs/btc_wrapper.log'),
        logging.StreamHandler()
    ]
)

class BTCWrapper:
    def __init__(self, config_path: str = 'config/wrapper_config.json'):
        """Initialize the BTC wrapper with configuration."""
        self.config = self._load_config(config_path)
        self.btc_rpc = self._init_bitcoin_rpc()
        self.web3 = self._init_web3()
        self.account = self._init_account()
        self.fernet = self._init_fernet()
        
    def _load_config(self, config_path: str) -> Dict:
        """Load configuration from file."""
        try:
            with open(config_path, 'r') as f:
                return json.load(f)
        except Exception as e:
            logging.error(f"Error loading config: {e}")
            sys.exit(1)
            
    def _init_bitcoin_rpc(self) -> RawProxy:
        """Initialize Bitcoin RPC connection."""
        try:
            return RawProxy(
                service_url=self.config['bitcoin_rpc_url'],
                service_port=self.config['bitcoin_rpc_port'],
                btc_conf_file=self.config['bitcoin_conf_path']
            )
        except Exception as e:
            logging.error(f"Error initializing Bitcoin RPC: {e}")
            sys.exit(1)
            
    def _init_web3(self) -> Web3:
        """Initialize Web3 connection."""
        try:
            w3 = Web3(Web3.HTTPProvider(self.config['k2_rpc_url']))
            if not w3.is_connected():
                raise Exception("Failed to connect to K2 network")
            return w3
        except Exception as e:
            logging.error(f"Error initializing Web3: {e}")
            sys.exit(1)
            
    def _init_account(self) -> Account:
        """Initialize Ethereum account for K2 transactions."""
        try:
            with open(self.config['key_file'], 'r') as f:
                encrypted_key = f.read()
            decrypted_key = self.fernet.decrypt(encrypted_key.encode())
            return Account.from_key(decrypted_key)
        except Exception as e:
            logging.error(f"Error initializing account: {e}")
            sys.exit(1)
            
    def _init_fernet(self) -> Fernet:
        """Initialize Fernet for key encryption."""
        try:
            with open(self.config['fernet_key_file'], 'r') as f:
                key = f.read().encode()
            return Fernet(key)
        except Exception as e:
            logging.error(f"Error initializing Fernet: {e}")
            sys.exit(1)
            
    async def wrap_rewards(self, amount: float, recipient: str) -> str:
        """Wrap Bitcoin rewards into K2 tokens."""
        try:
            # 1. Create Bitcoin transaction
            txid = self._create_bitcoin_tx(amount)
            logging.info(f"Created Bitcoin transaction: {txid}")
            
            # 2. Wait for confirmation
            await self._wait_for_confirmation(txid)
            
            # 3. Create K2 token minting transaction
            k2_tx_hash = await self._mint_k2_tokens(amount, recipient)
            logging.info(f"Minted K2 tokens: {k2_tx_hash}")
            
            return k2_tx_hash
        except Exception as e:
            logging.error(f"Error wrapping rewards: {e}")
            raise
            
    def _create_bitcoin_tx(self, amount: float) -> str:
        """Create a Bitcoin transaction to the wrapper address."""
        try:
            # Implementation details for Bitcoin transaction creation
            # This is a placeholder - actual implementation would use bitcoin.rpc
            return "mock_txid"
        except Exception as e:
            logging.error(f"Error creating Bitcoin transaction: {e}")
            raise
            
    async def _wait_for_confirmation(self, txid: str, confirmations: int = 6) -> None:
        """Wait for Bitcoin transaction confirmation."""
        try:
            # Implementation details for confirmation waiting
            # This is a placeholder - actual implementation would poll the RPC
            await asyncio.sleep(1)
        except Exception as e:
            logging.error(f"Error waiting for confirmation: {e}")
            raise
            
    async def _mint_k2_tokens(self, amount: float, recipient: str) -> str:
        """Mint K2 tokens for the recipient."""
        try:
            # Implementation details for K2 token minting
            # This is a placeholder - actual implementation would use web3
            return "mock_k2_tx_hash"
        except Exception as e:
            logging.error(f"Error minting K2 tokens: {e}")
            raise
            
    async def distribute_rewards(self, rewards: List[Dict[str, float]]) -> List[str]:
        """Distribute rewards to multiple recipients."""
        try:
            tx_hashes = []
            for reward in rewards:
                tx_hash = await self.wrap_rewards(
                    reward['amount'],
                    reward['recipient']
                )
                tx_hashes.append(tx_hash)
            return tx_hashes
        except Exception as e:
            logging.error(f"Error distributing rewards: {e}")
            raise

if __name__ == '__main__':
    wrapper = BTCWrapper()
    # Example usage
    rewards = [
        {'amount': 0.1, 'recipient': '0x123...'},
        {'amount': 0.2, 'recipient': '0x456...'}
    ]
    asyncio.run(wrapper.distribute_rewards(rewards)) 