phase-2/requirements.txt
flask==3.0.2
gunicorn==21.2.0
prometheus-client==0.19.0
psutil==5.9.8
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pytest-mock==3.12.0
pytest-asyncio==0.23.2
aiohttp==3.9.1
web3==6.11.1
eth-account==0.9.0
eth-typing==3.5.1
eth-utils==2.3.1
python-bitcoinlib==0.12.0
cryptography==41.0.7 